# auth-service/platform-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: IyziTrace Platform Gateway API
  version: 1.0.0
  summary: OpenAPI description of the public gateway endpoints exposed by the IyziTrace platform.
  description: |
    This specification documents the public HTTP surface exposed by the platform NGINX gateway.

    Included:
    - Platform core system status
    - OpAMP/Lawrence management APIs
    - Inventory APIs
    - OTLP and OpAMP ingestion routes
    - Curated query passthrough routes for metrics, logs, and traces

    Excluded:
    - `/api/v1/platform/auth/*` endpoints are intentionally omitted from this Swagger document.

    Notes:
    - Several endpoints are gateway passthroughs to upstream systems such as Thanos, Loki, Tempo, OTLP receivers, and Lawrence.
    - Authentication enforcement is configurable at runtime. The security schemes documented here describe the expected access pattern, but some routes may be open if the corresponding platform security toggle is disabled.
servers:
  - url: http://localhost
    description: Local HTTP gateway
  - url: https://localhost
    description: Local HTTPS gateway
...
paths:
...
  /ingest/otlp/v1/logs:
    post:
      tags: [Ingestion]
      operationId: ingestOtlpLogs
      summary: Ingest OTLP logs over HTTP
      description: |
        Gateway route for OpenTelemetry log ingestion. The request is forwarded to the log collector and rewritten to `/v1/logs`.
      security:
        - iyzitraceApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtlpJsonEnvelope"
            examples:
              sample:
                value:
                  resourceLogs:
                    - resource:
                        attributes:
                          - key: service.name
                            value:
                              stringValue: checkout
                      scopeLogs:
                        - logRecords:
                            - timeUnixNano: "1544712661000000000"
                              severityText: ERROR
                              body:
                                stringValue: Payment provider timeout
          application/x-protobuf:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Payload accepted by the upstream collector.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
        "400":
          description: Invalid OTLP payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Agent key required when ingress protection is enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/platform/opamp/topology:
    get:
      tags: [OpAMP Topology]
      operationId: getOpampTopology
      summary: Get overall Lawrence topology graph
      security:
        - bearerAuth: []
        - iyzitraceApiKey: []
      responses:
        "200":
          description: Topology returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopologyResponse"
...
components:
  securitySchemes:
    iyzitraceApiKey:
      type: apiKey
      in: header
      name: iyzitrace-api-key
      description: API key header used for agent ingress and external query consumers.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token accepted by platform middleware.
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    OtlpJsonEnvelope:
      type: object
      additionalProperties: true
      description: Simplified placeholder for OTLP JSON payloads.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
    NodeMetrics:
      type: object
      properties:
        metric_count:
          type: integer
          format: int64
        log_count:
          type: integer
          format: int64
        trace_count:
          type: integer
          format: int64
        error_rate:
          type: number
        latency:
          type: number
        throughput_rps:
          type: number
      required: [metric_count, log_count, trace_count, error_rate, latency, throughput_rps]
    TopologyNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        status:
          type: string
        group_id:
          type: string
          nullable: true
        group_name:
          type: string
          nullable: true
        labels:
          type: object
          additionalProperties:
            type: string
        metrics:
          $ref: "#/components/schemas/NodeMetrics"
        last_seen:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
      required: [id, type, name, status, labels]
    TopologyEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        type:
          type: string
        label:
          type: string
      required: [source, target, type]
    GroupSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agent_count:
          type: integer
        status:
          type: string
      required: [id, name, agent_count, status]
    TopologyResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/TopologyNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/TopologyEdge"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        services:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
      required: [nodes, edges, groups, services, updated_at]
...